Split execution bridge driver modules
