+++ .pm/tasks/task_41f5ad97727c4641b65241d4141601b9.execution.md
@@ -276,3 +276,19 @@ Example:
 - Review Findings Disposition: addressed
 - Finding Disposition Evidence: P1 simulator mirror durable-advance issue remains fixed in final HEAD by moving simulator world-dir persistence after expected peer execution hash validation and adding `rejected_expected_hash_does_not_persist_simulator_mirror`; P2 task metadata issue remains fixed by updated title/acceptance and project trace. Rebase conflict resolution preserved both runtime committed tick context and chain resource derivation context.
 - Residual Risk: wider full-suite runtime regression was not run; CI and live deployment/reset verification remain required before claiming testnet rebuilt.
+
+## 2026-06-22 15:33:00 CST / tpm
+- 完成内容: Fixed PR #558 required-gate failure after branch update.
+- Failure Signature: GitHub required-gate failed in `./scripts/check-rust-file-size.sh`; oversized files were `crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/driver.rs` at 1236 lines and `crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/tests/driver.rs` at 1370 lines, above the 1200-line limit.
+- Root Cause: current branch changes pushed both execution bridge production logic and tests over the repo file-size gate after rebase. Logic tests were green; the blocker was structural.
+- Implementation: extracted simulator mirror path/load/persist helpers into `execution_bridge/simulator_mirror.rs`; moved simulator payload-envelope test coverage into `tests/simulator_driver.rs`; moved storage profile/bootstrap driver tests into `tests/driver_storage_profile.rs`. Runtime behavior and assertions were preserved.
+- Validation Command: `./scripts/cargo-dev.sh fmt --all && git diff --check && ./scripts/check-rust-file-size.sh`
+- Expected Result: formatting succeeds, whitespace check passes, and no Rust file exceeds the 1200-line gate.
+- Actual Result: passed; `driver.rs` is 1194 lines and `tests/driver.rs` is 1107 lines; file-size gate reports oversized code/test files=0.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7 --bin oasis7_chain_runtime execution_bridge::tests:: -- --nocapture rejected_expected_hash_does_not_persist_simulator_mirror node_runtime_execution_driver_processes_simulator_payload_envelope production_release_policy_release_default_applies_hardened_policy new_driver_bootstraps_fresh_execution_world_persistence_files`
+- Expected Result: moved/split execution bridge tests compile and pass, including simulator rollback/persistence and storage-profile coverage.
+- Actual Result: passed; 50 execution_bridge tests ran and passed.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7_node synced_replication_commit_ -- --nocapture`
+- Expected Result: synced replication payload-node-id and rollback guardrail tests still pass after structural split.
+- Actual Result: passed; 2 tests ran and passed.
+- Residual Risk: CI required-gate must be rerun on the pushed fix; wider full-suite runtime regression remains unrun locally.
