+++ crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/driver.rs
@@ -37,6 +37,8 @@ use super::external_effect::{
     build_execution_external_effect_materialization,
     persist_execution_external_effect_materialization,
 };
+pub(crate) use super::simulator_mirror::simulator_world_dir_from_execution_world_dir;
+use super::simulator_mirror::{load_simulator_execution_world, persist_simulator_execution_world};
 use super::{
     EXECUTION_BRIDGE_DEFAULT_CHECKPOINT_INTERVAL_HEIGHTS,
     EXECUTION_BRIDGE_DEFAULT_CHECKPOINT_KEEP_LATEST, EXECUTION_BRIDGE_DEFAULT_HOT_WINDOW_HEIGHTS,
@@ -1046,50 +1048,6 @@ pub(crate) fn persist_execution_world(
     execution_world.save_to_dir(world_dir).map_err(|err| {
         format!(
             "save execution world to {} failed: {:?}",
-            world_dir.display(),
-            err
-        )
-    })
-}
-
-pub(crate) fn simulator_world_dir_from_execution_world_dir(world_dir: &Path) -> std::path::PathBuf {
-    match world_dir.file_name().and_then(|name| name.to_str()) {
-        Some(name) if !name.is_empty() => {
-            world_dir.with_file_name(format!("{name}-simulator-mirror"))
-        }
-        _ => world_dir.join("simulator-mirror"),
-    }
-}
-
-fn load_simulator_execution_world(world_dir: &Path) -> Result<WorldKernel, String> {
-    let snapshot_path = world_dir.join("snapshot.json");
-    let journal_path = world_dir.join("journal.json");
-    if !snapshot_path.exists() || !journal_path.exists() {
-        return Ok(WorldKernel::new());
-    }
-    WorldKernel::load_from_dir(world_dir).map_err(|err| {
-        format!(
-            "load simulator execution mirror from {} failed: {:?}",
-            world_dir.display(),
-            err
-        )
-    })
-}
-
-fn persist_simulator_execution_world(
-    world_dir: &Path,
-    simulator_world: &WorldKernel,
-    resource_context: Option<ChainResourceDerivationContext<'_>>,
-) -> Result<(), String> {
-    let result = match resource_context {
-        Some(context) => {
-            simulator_world.save_to_dir_with_chain_resource_context(world_dir, context)
-        }
-        None => simulator_world.save_to_dir(world_dir),
-    };
-    result.map_err(|err| {
-        format!(
-            "save simulator execution mirror to {} failed: {:?}",
             world_dir.display(),
             err
         )
