+++ crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/tests/driver.rs
@@ -17,10 +17,7 @@ use oasis7::runtime::{
     ModuleSubscriptionStage, ReleaseSecurityPolicy, WorldEventBody,
     production_hardened_main_token_config,
 };
-use oasis7::simulator::{
-    Action as SimulatorAction, ActionSubmitter, ChunkCoord, WorldConfig, WorldInitConfig,
-    WorldScenario, WorldSnapshot, initialize_kernel,
-};
+use oasis7::simulator::{Action as SimulatorAction, ActionSubmitter};
 use oasis7_node::{NodeExecutionCommitContext, NodeExecutionHook, compute_consensus_action_root};
 use oasis7_proto::storage_profile::StorageProfile;
 use oasis7_proto::storage_profile::StorageProfileConfig;
@@ -1106,265 +1103,5 @@ fn node_runtime_execution_driver_uses_storage_profile_checkpoint_interval() {
         Some(execution_checkpoint_manifest_rel_path(64).as_str())
     );
 
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn production_release_policy_release_default_applies_hardened_policy() {
-    let dir = temp_dir("execution-driver-release-policy-release-default");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let storage_profile = StorageProfileConfig::for_profile(StorageProfile::ReleaseDefault);
-    let driver = NodeRuntimeExecutionDriver::new_with_storage_profile(
-        state_path,
-        world_dir,
-        records_dir,
-        storage_root,
-        &storage_profile,
-    )
-    .expect("driver");
-
-    assert_eq!(
-        driver.execution_world.release_security_policy(),
-        &ReleaseSecurityPolicy::production_hardened()
-    );
-    assert!(
-        driver
-            .execution_world
-            .release_security_policy()
-            .is_production_hardened()
-    );
-
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn dev_local_storage_profile_keeps_generic_supply_for_missing_execution_world() {
-    let dir = temp_dir("execution-driver-release-policy-dev-local");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let storage_profile = StorageProfileConfig::for_profile(StorageProfile::DevLocal);
-    let driver = NodeRuntimeExecutionDriver::new_with_storage_profile(
-        state_path,
-        world_dir,
-        records_dir,
-        storage_root,
-        &storage_profile,
-    )
-    .expect("driver");
-
-    assert_eq!(
-        driver.execution_world.release_security_policy(),
-        &ReleaseSecurityPolicy::default()
-    );
-    assert_eq!(driver.execution_world.main_token_config().initial_supply, 0);
-
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn new_driver_bootstraps_fresh_execution_world_persistence_files() {
-    let dir = temp_dir("execution-driver-bootstrap-persistence");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let storage_profile = StorageProfileConfig::for_profile(StorageProfile::DevLocal);
-
-    let driver = NodeRuntimeExecutionDriver::new_with_storage_profile(
-        state_path,
-        world_dir.clone(),
-        records_dir,
-        storage_root,
-        &storage_profile,
-    )
-    .expect("driver");
-
-    assert!(world_dir.join("snapshot.json").exists());
-    assert!(world_dir.join("journal.json").exists());
-    assert!(driver.simulator_world_dir.join("snapshot.json").exists());
-    assert!(driver.simulator_world_dir.join("journal.json").exists());
-
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn new_driver_repairs_partial_execution_world_persistence_files() {
-    let dir = temp_dir("execution-driver-repair-partial-persistence");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let storage_profile = StorageProfileConfig::for_profile(StorageProfile::DevLocal);
-
-    fs::create_dir_all(world_dir.as_path()).expect("create world dir");
-    fs::write(world_dir.join("snapshot.json"), "{}").expect("write partial snapshot");
-
-    let driver = NodeRuntimeExecutionDriver::new_with_storage_profile(
-        state_path,
-        world_dir.clone(),
-        records_dir,
-        storage_root,
-        &storage_profile,
-    )
-    .expect("driver");
-
-    assert!(world_dir.join("snapshot.json").exists());
-    assert!(world_dir.join("journal.json").exists());
-    assert!(driver.simulator_world_dir.join("snapshot.json").exists());
-    assert!(driver.simulator_world_dir.join("journal.json").exists());
-
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn dev_local_storage_profile_clears_pristine_frozen_supply_from_existing_execution_world() {
-    let dir = temp_dir("execution-driver-release-policy-dev-local-existing-world");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let mut world = RuntimeWorld::new();
-    world.set_main_token_config(production_hardened_main_token_config());
-    persist_execution_world(world_dir.as_path(), &world).expect("persist release-like world");
-
-    let storage_profile = StorageProfileConfig::for_profile(StorageProfile::DevLocal);
-    let driver = NodeRuntimeExecutionDriver::new_with_storage_profile(
-        state_path,
-        world_dir,
-        records_dir,
-        storage_root,
-        &storage_profile,
-    )
-    .expect("driver");
-
-    assert_eq!(
-        driver.execution_world.release_security_policy(),
-        &ReleaseSecurityPolicy::default()
-    );
-    assert_eq!(driver.execution_world.main_token_config().initial_supply, 0);
-
-    let _ = fs::remove_dir_all(dir);
-}
-
-#[test]
-fn node_runtime_execution_driver_processes_simulator_payload_envelope() {
-    let dir = temp_dir("execution-driver-simulator-payload");
-    let state_path = dir.join("state.json");
-    let world_dir = dir.join("world");
-    let simulator_world_dir = simulator_world_dir_from_execution_world_dir(world_dir.as_path());
-    let records_dir = dir.join("records");
-    let storage_root = dir.join("store");
-    let mut driver = NodeRuntimeExecutionDriver::new(
-        state_path.clone(),
-        world_dir,
-        records_dir.clone(),
-        storage_root,
-    )
-    .expect("driver");
-    let config = WorldConfig::default();
-    let mut init =
-        WorldInitConfig::from_scenario(WorldScenario::AsteroidFragmentBootstrap, &config);
-    init.seed = 91;
-    init.asteroid_fragment.bootstrap_chunks = vec![ChunkCoord { x: 0, y: 0, z: 0 }];
-    let (kernel, _) = initialize_kernel(config, init).expect("simulator init");
-    driver.simulator_mirror = kernel;
-
-    let payload =
-        encode_consensus_action_payload(&ConsensusActionPayloadEnvelope::from_simulator_action(
-            SimulatorAction::HarvestRadiation {
-                agent_id: "agent-0".to_string(),
-                max_amount: 1,
-            },
-            ActionSubmitter::System,
-        ))
-        .expect("encode simulator payload");
-    let committed_action = oasis7_node::NodeConsensusAction::from_payload(1, "node-a", payload)
-        .expect("consensus action");
-    let action_root =
-        compute_consensus_action_root(std::slice::from_ref(&committed_action)).expect("root");
-    let expected_action_root = action_root.clone();
-    let expected_payload_hash = committed_action.payload_hash.clone();
-
-    let result = driver
-        .on_commit(NodeExecutionCommitContext {
-            world_id: "w1".to_string(),
-            node_id: "node-a".to_string(),
-            height: 1,
-            slot: 0,
-            epoch: 0,
-            node_block_hash: "node-h1".to_string(),
-            action_root,
-            committed_actions: vec![committed_action],
-            committed_at_unix_ms: 1_000,
-        })
-        .expect("commit");
-
-    assert_eq!(result.execution_height, 1);
-    assert!(records_dir.join("00000000000000000001.json").exists());
-    let record_bytes = fs::read(records_dir.join("00000000000000000001.json"))
-        .expect("read execution bridge record");
-    let record: ExecutionBridgeRecord =
-        serde_json::from_slice(record_bytes.as_slice()).expect("parse execution bridge record");
-    assert_eq!(record.schema_version, EXECUTION_BRIDGE_RECORD_SCHEMA_V2);
-    assert_eq!(
-        record.latest_state_ref.as_deref(),
-        record.snapshot_ref.as_deref()
-    );
-    assert!(
-        record
-            .snapshot_ref
-            .as_deref()
-            .is_some_and(|r| !r.is_empty())
-    );
-    assert!(record.journal_ref.as_deref().is_some_and(|r| !r.is_empty()));
-    let external_effect_ref = record
-        .external_effect_ref
-        .as_deref()
-        .expect("external effect ref should exist");
-    let store = LocalCasStore::new(dir.join("store"));
-    let external_effect =
-        load_execution_external_effect_materialization(&store, external_effect_ref)
-            .expect("load external effect materialization");
-    assert_eq!(external_effect.height, 1);
-    assert_eq!(external_effect.slot, 0);
-    assert_eq!(external_effect.epoch, 0);
-    assert_eq!(external_effect.action_root, expected_action_root);
-    assert_eq!(external_effect.committed_actions.len(), 1);
-    assert_eq!(external_effect.committed_actions[0].action_id, 1);
-    assert_eq!(
-        external_effect.committed_actions[0].payload_hash,
-        expected_payload_hash
-    );
-    assert!(external_effect.unresolved_inputs.is_empty());
-    let simulator = record
-        .simulator_mirror
-        .expect("simulator mirror record should exist");
-    assert_eq!(simulator.action_count, 1);
-    assert_eq!(simulator.rejected_action_count, 0);
-    assert!(!simulator.snapshot_ref.is_empty());
-    assert!(!simulator.journal_ref.is_empty());
-    assert!(!simulator.state_root.is_empty());
-    let simulator_snapshot_bytes = store
-        .get_verified(simulator.snapshot_ref.as_str())
-        .expect("load simulator snapshot");
-    let simulator_snapshot: WorldSnapshot =
-        serde_cbor::from_slice(simulator_snapshot_bytes.as_slice())
-            .expect("decode simulator snapshot");
-    let manifest = &simulator_snapshot.chain_resource_manifest;
-    let delta = &simulator_snapshot.latest_chain_resource_delta;
-    assert_eq!(
-        (manifest.world_id.as_str(), manifest.chain_id.as_str()),
-        ("w1", "w1")
-    );
-    assert_eq!(delta.commit_block_hash.as_deref(), Some("node-h1"));
-    assert!(!manifest.generated_chunks.is_empty());
-    assert!(!delta.entries.is_empty());
-    assert!(simulator_world_dir.join("snapshot.json").exists());
-    assert!(simulator_world_dir.join("journal.json").exists());
     let _ = fs::remove_dir_all(dir);
 }
